+++ docs/app/[...slug]/page.tsx
@@ -4,6 +4,7 @@ import { Metadata } from 'next';
 
 import { notFound } from 'next/navigation';
 
+import { TocContainer } from '@/ui/Toc';
 import { Mdx } from '@/ui/mdx';
 
 interface ContentPageProps {
@@ -45,10 +46,17 @@ export default async function ContentPage({ params }: ContentPageProps) {
   }
 
   return (
-    <article className="prose">
-      <Headline level={1}>{page.title}</Headline>
-      <div className="text-text-primary-muted -mt-8">{page.caption}</div>
-      <Mdx title={page.title} code={page.body.code} />
+    <article className="grid grid-cols-1 gap-x-24 gap-y-14 min-[1400px]:grid-cols-[minmax(min-content,70ch)_1fr]">
+      <div className="col-span-full">
+        <Headline level={1}>{page.title}</Headline>
+        <div className="text-secondary-400 pt-1">{page.caption}</div>
+      </div>
+      <div className="prose max-w-[70ch]">
+        <Mdx className="" title={page.title} code={page.body.code} />
+      </div>
+      <div className="col-start-2 hidden min-[1400px]:block">
+        <TocContainer />
+      </div>
     </article>
   );
 }