+++ docs/lib/config.ts
@@ -1,5 +1,6 @@
 export const siteConfig = {
   version: process.env.version,
+  defaultTheme: 'core',
   navigation: [
     {
       name: 'Introduction',
@@ -18,16 +19,8 @@ export const siteConfig = {
           slug: 'application',
         },
         {
-          name: 'Formatters',
-          slug: 'formatters',
-        },
-        {
           name: 'Layout',
           slug: 'layout',
-        },
-        {
-          name: 'Navigation',
-          slug: 'navigation',
         },
         {
           name: 'Form',
@@ -38,6 +31,10 @@ export const siteConfig = {
           slug: 'collection',
         },
         {
+          name: 'Navigation',
+          slug: 'navigation',
+        },
+        {
           name: 'Overlay',
           slug: 'overlay',
         },
@@ -46,10 +43,18 @@ export const siteConfig = {
           slug: 'content',
         },
         {
+          name: 'Formatters',
+          slug: 'formatters',
+        },
+        {
           name: 'Hooks and Utils',
           slug: 'hooks-and-utils',
         },
       ],
+    },
+    {
+      name: 'Patterns',
+      slug: 'patterns',
     },
     {
       name: 'Recipes',