+++ packages/components/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @marigold/components
 
+## 7.7.0
+
+### Minor Changes
+
+- [#3892](https://github.com/marigold-ui/marigold/pull/3892) [`3f02ea1`](https://github.com/marigold-ui/marigold/commit/3f02ea15fab7b27907b0b478d479d6f2766e3ab2) Thanks [@sebald](https://github.com/sebald)! - feat: Allow to disable keyboard navigation on `Table`
+
+### Patch Changes
+
+- [#3885](https://github.com/marigold-ui/marigold/pull/3885) [`72ece08`](https://github.com/marigold-ui/marigold/commit/72ece08fe8009ee19b05c2ad8796658dfa91ebb8) Thanks [@sebald](https://github.com/sebald)! - fix: `getColor` correctly works with deeply nested values
+
+- [#3880](https://github.com/marigold-ui/marigold/pull/3880) [`4a59427`](https://github.com/marigold-ui/marigold/commit/4a59427fc1d4c14b1971c07778de1977cdde5dda) Thanks [@sebald](https://github.com/sebald)! - chore: cleanup peerdeeps
+
+- Updated dependencies [[`72ece08`](https://github.com/marigold-ui/marigold/commit/72ece08fe8009ee19b05c2ad8796658dfa91ebb8), [`4a59427`](https://github.com/marigold-ui/marigold/commit/4a59427fc1d4c14b1971c07778de1977cdde5dda)]:
+  - @marigold/system@7.7.0
+  - @marigold/icons@1.2.46
+
 ## 7.6.0
 
 ### Minor Changes